Modify route protection logic
